pyproject.toml
[project]
name = "uicdev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-jazzmin>=3.0.4",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.29.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.2",
]
